Fix UITableView issue caused by iOS 11 update (Seriously???)
When programmatically reloading & scrolling at the same time, the scroll animation will be interrupted by tableview reload. This issue is introduced in iOS 11 and still exists in current iOS version (12). Reloading & user interaction scrolling will not trigger this issue. Solution is adding a 0 second animation block to reloading and scrolling the tableview in completion block.